requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
deep-translator==1.11.4
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
gTTS==2.5.1 